Restuctured and fixed tests
